Generalize counting operation in ecosystem
Also switch to using constants instead of magic labels for Thing
subtypes.

+++ src/main/java/org/nusco/narjillos/experiment/environment/ThingsCounter.java
@@ -5,23 +5,23 @@ import java.util.Map;
 
 class ThingsCounter {
 
-	private final Map<String, Integer> countsByLabel = new HashMap<>();
+	private final Map<String, Long> countsByLabel = new HashMap<>();
 
 	public synchronized void add(String label) {
 		if (countsByLabel.containsKey(label))
 			countsByLabel.put(label, countsByLabel.get(label) + 1);
 		else
-			countsByLabel.put(label, 1);
+			countsByLabel.put(label, 1L);
 	}
 
 	public synchronized void remove(String label) {
 		countsByLabel.put(label, countsByLabel.get(label) - 1);
 	}
 
-	public synchronized int count(String label) {
-		Integer result = countsByLabel.get(label);
+	public synchronized long count(String label) {
+		Long result = countsByLabel.get(label);
 		if (result == null)
-			return 0;
+			return 0L;
 		return result;
 	}
 }
